requirements.txt
google-cloud-speech==0.29.0